Fix RUF067: move compile_program and register_dialects out of __init__.py
ruff RUF067 requires __init__ modules to only contain docstrings and
re-exports. Move the function definitions to _pipeline.py and re-export
from __init__.py.
